+ synchronous MDN processing

+++ src/org/redoubt/protocol/as2/As2HttpListener.java
@@ -63,8 +63,10 @@ public class As2HttpListener extends HttpServlet {
             sLogger.error("An error has occured while processing inbound AS2 message. " + e.getMessage(), e);
             throw new IOException(e.getMessage(), e);
         } finally {
-            FileSystemUtils.backupFile(workFile);
-            FileSystemUtils.removeWorkFile(workFile);
+        	if(workFile != null) {
+	            FileSystemUtils.backupFile(workFile);
+	            FileSystemUtils.removeWorkFile(workFile);
+        	}
         }
         
         prepareMdnResponse(context, resp);
